feat: set up env.d.ts on `astro sync`
